pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "news-fetcher"
version = "0.1.0"
authors = [
  { name = "News Fetcher Team", email = "contact@newsfetcher.example" }
]
description = "A news aggregation and clustering tool that fetches, processes, and clusters news articles from various sources."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests>=2.31.0",
    "feedparser>=6.0.11",
    "beautifulsoup4>=4.12.2",
    "numpy>=1.24.3",
    "scikit-learn>=1.3.0",
    "nltk>=3.8.1",
    "pyyaml>=6.0.1",
    "click>=8.1.3"
]

[project.urls]
"Homepage" = "https://github.com/example/news-fetcher"
"Bug Tracker" = "https://github.com/example/news-fetcher/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "responses>=0.23.3",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

[project.scripts]
news-fetcher = "news_fetcher.cli:main"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | __pycache__
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
strict = true
ignore_missing_imports = false
